requirements.txt
anyascii==0.3.0
appdirs==1.4.3
asgiref==3.5.0
attrs==20.3.0
beautifulsoup4==4.9.3
boto3==1.17.37
botocore==1.20.37
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
distro==1.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-cors-headers==3.7.0
django-createsuperuser==2020.12.3
django-filter==21.1
django-heroku==0.3.1
django-modelcluster==5.2
django-storages==1.11.1
django-taggit==1.5.1
django-toolbelt==0.0.1
django-treebeard==4.5.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
draftjs-exporter==2.1.7
drf-yasg==1.20.0
et-xmlfile==1.1.0
filelock==3.4.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
l18n==2021.3
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
openpyxl==3.0.9
packaging==20.3
pep517==0.8.2
Pillow==8.1.2
pipenv==2020.11.15
platformdirs==2.4.1
progress==1.5
psycopg2==2.8.6
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
six==1.14.0
soupsieve==2.3.1
sqlparse==0.4.1
static3==0.7.0
swagger-spec-validator==2.7.3
tablib==3.2.0
telepath==0.2
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.13.0
virtualenv-clone==0.5.4
wagtail==2.15.3
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
